110623 mañana

+++ src/main/java/items/GreatSword.java
@@ -4,7 +4,7 @@ import source.Weapon;
 
 public class GreatSword extends Weapon{
 
-	public GreatSword(String name, int tier, int damageMin, int damageMax, double critMod, int speedMod) {
+	public GreatSword(String name, int tier, int damageMin, int damageMax, int critMod, int speedMod) {
 		super(name, tier, damageMin, damageMax, critMod, speedMod);
 	}
