+++ src/main/java/items/Mace.java
@@ -4,7 +4,7 @@ import source.Weapon;
 
 public class Mace extends Weapon {
 
-	public Mace(String name, int tier, int damageMin, int damageMax, double critMod, int speedMod) {
+	public Mace(String name, int tier, int damageMin, int damageMax, int critMod, int speedMod) {
 		super(name, tier, damageMin, damageMax, critMod, speedMod);
 	}
